requirements.txt
a2a-sdk==0.3.20
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiomysql==0.3.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
asyncpg==0.31.0
attrs==25.4.0
aws-requests-auth==0.4.3
awscli==1.43.10
beautifulsoup4==4.14.3
boto3==1.42.4
botocore==1.42.4
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dill==0.4.0
docstring_parser==0.17.0
docutils==0.19
exceptiongroup==1.3.1
fastapi==0.124.0
frozenlist==1.8.0
google-api-core==2.28.1
google-auth==2.43.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
markdownify==1.2.2
mcp==1.23.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
numpy==2.2.6
opentelemetry-api==1.39.0
opentelemetry-instrumentation==0.60b0
opentelemetry-instrumentation-threading==0.60b0
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
rsa==4.7.2
s3transfer==0.16.0
six==1.17.0
slack_bolt==1.27.0
slack_sdk==3.39.0
soupsieve==2.8
SQLAlchemy==2.0.44
sse-starlette==3.0.3
starlette==0.50.0
strands-agents==1.19.0
strands-agents-tools==0.2.17
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
uvicorn==0.38.0
watchdog==6.0.0
wcwidth==0.2.14
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
